Store DPI scale as a float instead of a double

+++ include/video.h
@@ -414,7 +414,7 @@ DosBox::Rect GFX_GetCanvasSizeInPixels();
 DosBox::Rect GFX_GetViewportSizeInPixels();
 DosBox::Rect GFX_GetDesktopSize();
 
-double GFX_GetDpiScaleFactor();
+float GFX_GetDpiScaleFactor();
 
 RenderingBackend GFX_GetRenderingBackend();
